Responses now successful unless error explicit
Fileuploader now looks for explicit "error" field to determine if there
is an API level error (after check for HTTP error) instead of forcing
API to send "success" field in JSON. That's just silly and forces
library to not work out of the box for many popular RESTful APIs.

We still add an explicit error field, which is still wrong, it just
made changes faster. We could also check for "status": false, but that
continues to violate the point that the client should check these
things based on the API.

There are IE bugs that prevent proper implementations of such a design.
Screw IE.